Split config into a distinct module for each component.
